fix(wasm): opfs project (#2109)
